requirements.txt
dotenv
psycopg2-binary
fastapi[standard]
uvicorn
pydantic